clean arch refactor
