+++ src/main/java/br/com/gearing/guru/adapters/api/ExternalService.java
@@ -1,0 +1,4 @@
+package br.com.gearing.guru.adapters.api;
+
+public class ExternalService {
+}
