+++ src/main/java/br/com/gearing/guru/adapters/repository/CalculationRepository.java
@@ -1,0 +1,4 @@
+package br.com.gearing.guru.adapters.repository;
+
+public interface CalculationRepository {
+}
