+++ src/main/java/br/com/gearing/guru/adapters/web/CalcController.java
@@ -1,0 +1,4 @@
+package br.com.gearing.guru.adapters.web;
+
+public class CalcController {
+}
