--- src/main/java/br/com/gearing/guru/api/controllers/HealthController.java
@@ -1,18 +1,0 @@
-package br.com.gearing.guru.api.controllers;
-
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HealthController {
-
-    @GetMapping("/health")
-    public ResponseEntity<String> healthStatus() {
-        return new ResponseEntity<>(HttpStatus.OK);
-
-    }
-
-}
