+++ src/main/java/br/com/gearing/guru/application/CalculationUseCase.java
@@ -1,0 +1,4 @@
+package br.com.gearing.guru.application;
+
+public class CalculationUseCase {
+}
