--- src/main/java/br/com/gearing/guru/domain/entities/UserEntity.java
@@ -1,32 +1,0 @@
-package br.com.gearing.guru.domain.entities;
-
-
-import jakarta.persistence.*;
-import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
-
-@Data
-@Getter
-@Setter
-@Table(schema = "users")
-public class UserEntity {
-
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @Column(name = "name")
-    private String userName;
-    @Column(name = "email")
-    private String email;
-    @Column(name = "password")
-    private String password;
-
-
-    public UserEntity(String userName, String email, String password) {
-        this.userName = userName;
-        this.email = email;
-        this.password = password;
-    }
-}
