+++ src/main/java/br/com/gearing/guru/domain/model/Calculation.java
@@ -1,0 +1,4 @@
+package br.com.gearing.guru.domain.model;
+
+public class Calculation {
+}
