+++ src/main/java/br/com/gearing/guru/infrastructure/GearingGuruApplication.java
@@ -1,4 +1,4 @@
-package br.com.gearing.guru.application;
+package br.com.gearing.guru.infrastructure;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
