--- src/main/java/br/com/gearing/guru/infrastructure/handler/UserCommandsHandler.java
@@ -1,10 +1,0 @@
-package br.com.gearing.guru.infrastructure.handler;
-
-
-import org.springframework.web.bind.annotation.ControllerAdvice;
-
-@ControllerAdvice("br.com.gearing.guru.api.controllers")
-public class UserCommandsHandler {
-
-
-}
